Armstrong number Checker
This program can check if a number is an Armstrong number. An arm strong number is that is equal to the sum of its digits, each raised to a power.
We can use a while loop to check this. We also use the cmath library to to calculate exponential terms instead of just multiplying again and again.